pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-examples"
version = "0.1.0"
description = "Google ADK Agent Examples"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
    "mcp>=1.1.3",
    "robin-stocks",
    "pydantic>=2.0.0",
    "hatch-vcs",
    "anthropic>=0.64.0",
    "strands-agents[anthropic]>=1.1.0",
    "strands-agents-tools>=0.2.4",
    "python-dotenv>=1.1.1",
    "pandas>=2.3.2",
    "numpy>=2.2.6",
    "openai>=1.101.0",
    "litellm>=1.75.9",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "basic-open-agent-tools>=0.8.2",
    "requests>=2.32.4",
    "httpx>=0.28.1",
    "aiohttp>=3.12.15",
    "rich>=14.1.0",
    "click>=8.2.1",
    "typer>=0.16.1",
    "openpyxl>=3.1.5",
    "markdown>=3.8.2",
    "pyyaml>=6.0.2",
    "google-adk>=1.10.0",
    "pytest>=8.4.1",
    "tavily-python>=0.7.11",
]

[tool.hatch.build.targets.wheel]
packages = ["GoogleADK", "scripts", "AWS_Strands.strands_docs_mcp-server.src.strands_mcp_server"]
